added flags l and ll for d,o,u conversins, added ft_uitoa_base function for u conversion

+++ libft/ft_itoa.c
@@ -6,13 +6,13 @@
 /*   By: vhazelnu <vhazelnu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/21 14:28:47 by vhazelnu          #+#    #+#             */
-/*   Updated: 2019/06/02 11:07:02 by vhazelnu         ###   ########.fr       */
+/*   Updated: 2019/06/04 09:56:53 by vhazelnu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int			ft_count_digit(int n, int len)
+int			ft_count_digit(long long int n, int len)
 {
 	while (n / 10)
 	{
@@ -22,11 +22,11 @@ int			ft_count_digit(int n, int len)
 	return (len);
 }
 
-char		*ft_itoa(int n)
+char		*ft_itoa(long long int n)
 {
-	char			*result;
-	int				len;
-	unsigned int	ui;
+	char					*result;
+	int						len;
+	unsigned long long int	ui;
 
 	len = n < 0 ? 2 : 1;
 	len = ft_count_digit(n, len);
